feat: egress 하네스에 EGRESS_WIDTH/HEIGHT 해상도 knob 추가

+++ internal/media/egress_audio_harness_test.go
@@ -121,8 +121,8 @@ feed:
 				data:      harnessFrameData(t, index, harnessWireFormat()),
 				timestamp: timestamp,
 				stageAt:   time.Now(),
-				width:     harnessWidth,
-				height:    harnessHeight,
+				width:     uint16(harnessWidth),
+				height:    uint16(harnessHeight),
 			})
 			timestamp += step
 			index++
