+++ internal/media/egress_integration_test.go
@@ -72,8 +72,8 @@ feed:
 			egress.Enqueue(frame{
 				data:      harnessJPEG(t, index),
 				timestamp: timestamp,
-				width:     harnessWidth,
-				height:    harnessHeight,
+				width:     uint16(harnessWidth),
+				height:    uint16(harnessHeight),
 			})
 			timestamp += step
 			index++
